9 спринт. Исправление замечаний ревьювера.
