+++ src/main/HTTPserver/PrioritizedHandler.java
@@ -1,0 +1,18 @@
+package main.HTTPserver;
+
+import com.sun.net.httpserver.HttpExchange;
+import com.sun.net.httpserver.HttpHandler;
+
+import java.io.IOException;
+
+import static main.HTTPserver.BaseHttp.*;
+import static main.HTTPserver.HttpTaskServer.manager;
+
+// метод обработки ветки /prioritized
+class PrioritizedHandler implements HttpHandler {
+    @Override
+    public void handle(HttpExchange exchange) throws IOException {
+        sendResponse(exchange, toJsonString(manager.getPrioritizedTasks()), 200);
+    }
+}
+
