+++ src/main/Main.java
@@ -13,7 +13,7 @@ import java.io.IOException;
 import java.time.Duration;
 import java.time.LocalDateTime;
 
-public class TestMain {
+public class Main {
 
     public static void main(String[] args) {
