+++ src/main/classes/Status.java
@@ -1,5 +1,7 @@
 package main.classes;// Задаём параметры состояний задач
 
 public enum Status {
-    NEW, IN_PROGRESS, DONE
+    NEW,
+    IN_PROGRESS,
+    DONE
 }
