+++ test/CheckWithFile.java
@@ -4,16 +4,19 @@ import main.classes.Epic;
 import main.classes.Status;
 import main.classes.SubTask;
 import main.classes.Task;
+import main.service.ManagerSaveException;
 import main.service.Managers;
 import main.service.TaskManager;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import java.io.File;
+import java.io.IOException;
 import java.time.Duration;
 import java.time.LocalDateTime;
 
-class TestInMemory {
+class CheckWithFile {
 
     private static TaskManager testManager;
     private static Task taskItem;
@@ -22,18 +25,90 @@ class TestInMemory {
 
     @BeforeEach
     void initializeTaskManager() {
+        // Создаем пустой файл для менеджера задач
+        File testFile;
+        try {
+            testFile = File.createTempFile("test", ".csv");
+        } catch (IOException e) {
+            throw new ManagerSaveException("Не удалось создать временный файл");
+        }
+
+        // Проверяем что файл создан и он пустой
+        Assertions.assertNotNull(testFile);
+        // Проверяем что файл пустой
+        Assertions.assertEquals(0, testFile.length());
+
         // Проверяем создание менеджера через утилитарный класс, используем его в тестах
-        testManager = Managers.getDefault();
+        testManager = Managers.getFile(testFile);
+
         Assertions.assertNotNull(testManager); // проверяем что менеджер создан
         Assertions.assertNotNull(testManager.getAllTasks()); // проверяем что список задач создан
         Assertions.assertNotNull(testManager.getAllEpics()); // проверяем что список эпиков создан
         Assertions.assertNotNull(testManager.getAllSubTasks()); // проверяем что список подзадач создан
         Assertions.assertNotNull(testManager.getHistory()); // проверяем что история создана
 
+        // Если при создании менеджера не было ошибок - значит пустой файл был считан и тест пройден.
+
+    }
+
+    @Test
+    void WriteAndReadDataFile_CorrectInputOutput() {
+        // Проверяем что менеджер сохраняет данные в файл
+
         // Создаём элементы для добавления в менеджер
         taskItem = new Task(0,
                 "Работа", "Просто задача", Status.NEW,
                 LocalDateTime.of(2024, 1, 1, 10, 5, 0),
+                Duration.ofMinutes(50));
+        epicItem = new Epic(0,
+                "Этапы", "Поэтапная работа", Status.NEW);
+        subTaskItem = new SubTask(0,
+                "Этап 1", "Подготовка к работе", Status.NEW, 2,
+                LocalDateTime.of(2024, 1, 5, 8, 5, 0),
+                Duration.ofMinutes(250));
+
+        // Добавляем элементы в менеджер
+        testManager.addTask(taskItem);
+        testManager.addEpic(epicItem);
+        testManager.addSubTask(subTaskItem);
+
+        // Проверяем что файл не пустой
+        Assertions.assertNotEquals(0, testManager.getFile().length());
+
+        // Считываем данные из файла в другой экземпляр менеджера
+        TaskManager testManager2 = Managers.getFile(testManager.getFile());
+
+        // Проверяем что данные из файла были считаны правильно и равны исходным
+        Assertions.assertEquals(testManager.getAllTasks(), testManager2.getAllTasks());
+        Assertions.assertEquals(testManager.getAllEpics(), testManager2.getAllEpics());
+        Assertions.assertEquals(testManager.getAllSubTasks(), testManager2.getAllSubTasks());
+        // Проверяем идентичность полученные элементы из файла с исходными по всем значениям
+        Assertions.assertEquals(taskItem, testManager2.getTask(taskItem.getId()));
+        Assertions.assertEquals(epicItem, testManager2.getEpic(epicItem.getId()));
+        Assertions.assertEquals(subTaskItem, testManager2.getSubTask(subTaskItem.getId()));
+
+        // Пройденные тесты говорят о правильности чтения из файла.
+
+        // Удаляем содержимое второго экземпляра менеджера
+        testManager2.deleteAllSubTasks();
+        testManager2.deleteAllEpics();
+        testManager2.deleteAllTasks();
+        // Проверяем что менеджер пустой
+        Assertions.assertTrue(testManager2.getAllTasks().isEmpty());
+        Assertions.assertTrue(testManager2.getAllEpics().isEmpty());
+        Assertions.assertTrue(testManager2.getAllSubTasks().isEmpty());
+        // Проверяем, что файл второго менеджера пустой
+        Assertions.assertEquals(0, testManager2.getFile().length());
+
+    }
+
+    @Test
+    void AddItem_CorrectOfRecordInputAndSave() {
+
+        // Создаём элементы для добавления в менеджер
+        taskItem = new Task(0,
+                "Работа", "Просто задача", Status.NEW,
+                LocalDateTime.of(2024, 1, 5, 10, 5, 0),
                 Duration.ofMinutes(50));
         epicItem = new Epic(0,
                 "Этапы", "Поэтапная работа", Status.NEW);
@@ -46,11 +121,6 @@ class TestInMemory {
         testManager.addTask(taskItem);
         testManager.addEpic(epicItem);
         testManager.addSubTask(subTaskItem);
-
-    }
-
-    @Test
-    void addItem() {
 
         // Проверяем, что globalId имеет последний свободный id номер
         Assertions.assertEquals(4, testManager.getGlobalId());
@@ -90,7 +160,25 @@ class TestInMemory {
 
 
     @Test
-    void testItems() {
+    void checkIdentityOfItemsWithOneId() {
+
+        // Создаём элементы для добавления в менеджер
+        taskItem = new Task(0,
+                "Работа", "Просто задача", Status.NEW,
+                LocalDateTime.of(2024, 1, 1, 10, 5, 0),
+                Duration.ofMinutes(50));
+        epicItem = new Epic(0,
+                "Этапы", "Поэтапная работа", Status.NEW);
+        subTaskItem = new SubTask(0,
+                "Этап 1", "Подготовка к работе", Status.NEW, 2,
+                LocalDateTime.of(2024, 1, 1, 10, 5, 0),
+                Duration.ofMinutes(250));
+
+        // Добавляем элементы в менеджер
+        testManager.addTask(taskItem);
+        testManager.addEpic(epicItem);
+        testManager.addSubTask(subTaskItem);
+
         // проверяем равенство элементов с одним и тем-же id
         Task taskItem1 = testManager.getTask(1);
         Task taskItem2 = testManager.getTask(1);
@@ -105,9 +193,27 @@ class TestInMemory {
     }
 
     @Test
-    void testHistory1() {
-        // Проверяем, что в истории только 1 запись - стартовая
-        Assertions.assertEquals(1, testManager.getHistory().size(), "В истории не одна запись.");
+    void AddToHistoryAndChanges_HistoryIsCreatedRecordedCorrectlyAndChanged() {
+
+        // Создаём элементы для добавления в менеджер
+        taskItem = new Task(0,
+                "Работа", "Просто задача", Status.NEW,
+                LocalDateTime.of(2024, 1, 1, 12, 5, 0),
+                Duration.ofMinutes(50));
+        epicItem = new Epic(0,
+                "Этапы", "Поэтапная работа", Status.NEW);
+        subTaskItem = new SubTask(0,
+                "Этап 1", "Подготовка к работе", Status.NEW, 2,
+                LocalDateTime.of(2024, 1, 1, 8, 5, 0),
+                Duration.ofMinutes(250));
+
+        // Добавляем элементы в менеджер
+        testManager.addTask(taskItem);
+        testManager.addEpic(epicItem);
+        testManager.addSubTask(subTaskItem);
+
+        // Проверяем, что в истории только стартовая запись
+        Assertions.assertEquals(1, testManager.getHistory().size(), "История не пустая.");
         // Добавляем элемент в историю при помощи запроса к менеджеру
         Task taskItem1 = testManager.getTask(1);
         // проверяем, что история изменилась на 1
@@ -135,10 +241,27 @@ class TestInMemory {
     }
 
     @Test
-    void testHistory2() {
+    void ChangesInHistory_DeletingItem_historyIsDeletedAndNoRepeatHistory() {
+
+        // Создаём элементы для добавления в менеджер
+        taskItem = new Task(0,
+                "Работа", "Просто задача", Status.NEW,
+                LocalDateTime.of(2024, 3, 1, 10, 5, 0),
+                Duration.ofMinutes(50));
+        epicItem = new Epic(0,
+                "Этапы", "Поэтапная работа", Status.NEW);
+        subTaskItem = new SubTask(0,
+                "Этап 1", "Подготовка к работе", Status.NEW, 2,
+                LocalDateTime.of(2024, 1, 1, 8, 5, 0),
+                Duration.ofMinutes(250));
+
+        // Добавляем элементы в менеджер
+        testManager.addTask(taskItem);
+        testManager.addEpic(epicItem);
+        testManager.addSubTask(subTaskItem);
+
         // Проверяем, что в истории только стартовая запись
-        Assertions.assertEquals(1,
-                testManager.getHistory().size(), "В истории не одна запись.");
+        Assertions.assertEquals(1, testManager.getHistory().size(), "История не пустая.");
         // Добавляем 1 элемент в историю при помощи запроса к менеджеру
         Task taskItem1 = testManager.getTask(1);
         // проверяем, что история изменилась на 1
@@ -153,7 +276,7 @@ class TestInMemory {
         taskItem1 = testManager.getTask(1);
         taskItem1 = testManager.getSubTask(3);
         taskItem1 = testManager.getEpic(2);
-        // Проверяем, что история осталась с 4 элементам. Что дублирования записей нет.
+        // Проверяем, что история осталась с 3 элементам. Что дублирования записей нет.
         Assertions.assertEquals(3,
                 testManager.getHistory().size(), "В истории появились дубли.");
         // удаляем задачу и проверяем состояние истории - уменьшилась ли она, и есть ли запись с удалённой задачей.
